fixed int/double mapper output bug

+++ src/main/java/AverageRatingReduceClass.java
@@ -1,19 +1,18 @@
 import java.io.IOException;
 
 import org.apache.hadoop.io.DoubleWritable;
-import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Reducer;
 
-public class AverageRatingReduceClass extends Reducer<Text,IntWritable,Text,DoubleWritable> {
+public class AverageRatingReduceClass extends Reducer<Text,DoubleWritable,Text,DoubleWritable> {
 
     @Override
-    protected void reduce(Text key, Iterable<IntWritable> values, Context context)
+    protected void reduce(Text key, Iterable<DoubleWritable> values, Context context)
             throws IOException, InterruptedException {
 
         double sum = 0;
         double totalReviews = 0;
-        for (IntWritable val : values) {
+        for (DoubleWritable val : values) {
             sum += val.get();
             ++totalReviews;
         }
